break receiver stop section into component
